pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lenia_field"
version = "0.1.0"
description = "Real-time Lenia morphogen field with external stimulus injection for bio-hybrid systems"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Charles" }
]
keywords = [
    "lenia",
    "cellular-automata",
    "morphogen",
    "bio-hybrid",
    "artificial-life",
    "jax",
    "real-time"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Life",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "jax>=0.4.0",
    "jaxlib>=0.4.0",
    "numpy>=1.20.0",
    "pyzmq>=22.0.0",
    "msgpack>=1.0.0",
    "msgpack-numpy>=0.4.8",
]

[project.optional-dependencies]
cuda11 = [
    "jax[cuda11_pip]>=0.4.0",
]
cuda12 = [
    "jax[cuda12]>=0.4.0",
]
viz = [
    "opencv-python>=4.5.0",
    "matplotlib>=3.5.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-benchmark>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "lenia_field[viz,dev]",
]

[project.scripts]
lenia-server = "lenia_field.server:main"

[project.urls]
Homepage = "https://github.com/yourusername/lenia_field"
Documentation = "https://github.com/yourusername/lenia_field#readme"
Repository = "https://github.com/yourusername/lenia_field"
Issues = "https://github.com/yourusername/lenia_field/issues"

[tool.hatch.build.targets.wheel]
packages = ["lenia_field"]

[tool.hatch.build.targets.sdist]
include = [
    "/lenia_field",
    "/README.md",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # function calls in argument defaults
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
